Fix the capturing of IPv6 UDP packets
In the while { } loop in main() there is a check to see if the current
packet is UDP or TCP. Unfortunately the check only applies to IPv4 and
any IPv6 packets will be classed as TCP regardless.

This then causes the subsequent data variable which we look in for the
SIP method etc to be somewhat truncated (due to now incorrect offsets),
generally leading to no valid SIP method being found and the packet not
being captured.

To check IPv6 packets we need to look at the 'Next Header' field of the
IPv6 packet.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>

git-svn-id: https://svn.code.sf.net/p/pcapsipdump/code/trunk@135 89f3ae11-3634-4b71-aec2-a73e79b8c089